fix: [tracing docs] make Jaeger and OpenTelemetry work
Recent releases of OpenTelemetry collector assume the connection between
the collector and the GRPC endpoint is encrypted via TLS.

That's not the case in our simplified deployment. Because of that, we
have to provide some extra information to the end users to ensure the
communication between the OpenTelemetry collector and the Jaeger
endpoint works.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>