refactor: small tweaks to monitor mode docs
Apparently we though about writing multiple documents about the monitor
mode, but we never did. Because of that we had a dedicated folder with
just one single markdown file called "intro".

The current one seems good, hence we can get rid of this filesystem
layout and, most important to the end user, pick a better name for the
page.

Overall changes:

* Change the sidebar title to be more concise
* On the filesystem, get rid of the folder and have just one single
markdown file

Signed-off-by: Flavio Castelli <fcastelli@suse.com>