Sidebar revamp based on divio paradigm
As for the other SUSE Rancher projects, the Divio paradigm will be implemented for Kubewarden docs.

This change will be done in different steps and should bring an easier navigation for the docs.

This first step is a complete rework on the sidebar, to align it with the Divio paradigm.

The docs will be re-arranged under the four main groups, however if needed, additional groups will be created.

Signed-off-by: Nuno do Carmo <nuno.carmo@suse.com>