Signed-off-by: John Krug <john.krug@suse.com>
Added a reminder for making sure you have a GH access token for ghcr.

Added some metadata for SEo to frontmatter, description and keywords.

Refreshed some language, trying to make it all a little crisper. Mainly
by removing words. Especially for a quick-start the less to navigate and
digest the better.

Some tables had a mix of of unicode ticks and text dashes to indicate
Yes or No. Standardised on 'Y' and 'N'. Happy to discuss this further.