fix: group policies CRD definition was wrong
The CRDs of the policy group were probably generated a bit earlier
during the development cycle. In the beginning the policies part of the
group were kept inside of an array, but later we moved to use a
dictionary.

The documentation we shipped made reference to the array of policies,
instead of the dictionary.

This commit fixes the issue also on the older releases, since we never
ever shipped the CRD that used the list.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>